requirements.txt
pandas>=0.21.1
google-cloud-translate>=1.3.0
